Delete evil files, with API calls :)
Fix the issue with the API calls. Now the program actually checks
the response from API and delete the file if needs.

curl.h: the downloaded content may have '\0's, until better fix, just
replace them with something.
TODO: This is still an issue, the `strstr` function may have not find
the strings if there are external characters inserted.